Fixed warnings in semaphore code as suggested in issue #108

+++ atomicops.h
@@ -460,11 +460,11 @@ namespace moodycamel
 				return timed_wait(0);
 			}
 
-			bool timed_wait(std::int64_t timeout_usecs) AE_NO_TSAN
+			bool timed_wait(std::uint64_t timeout_usecs) AE_NO_TSAN
 			{
 				mach_timespec_t ts;
 				ts.tv_sec = static_cast<unsigned int>(timeout_usecs / 1000000);
-				ts.tv_nsec = (timeout_usecs % 1000000) * 1000;
+				ts.tv_nsec = static_cast<int>((timeout_usecs % 1000000) * 1000);
 
 				// added in OSX 10.10: https://developer.apple.com/library/prerelease/mac/documentation/General/Reference/APIDiffsMacOSX10_10SeedDiff/modules/Darwin.html
 				kern_return_t rc = semaphore_timedwait(m_sema, ts);
@@ -500,7 +500,7 @@ namespace moodycamel
 		    AE_NO_TSAN Semaphore(int initialCount = 0)
 		    {
 		        assert(initialCount >= 0);
-		        int rc = sem_init(&m_sema, 0, initialCount);
+		        int rc = sem_init(&m_sema, 0, static_cast<unsigned int>(initialCount));
 		        assert(rc == 0);
 		        AE_UNUSED(rc);
 		    }
